1.修改页面ui bug

+++ VERSION1.0/ClientEngine/UI/src/NDUILabel.cpp
@@ -123,7 +123,7 @@ void NDUILabel::SetFontColor(ccColor4B fontColor)
 	
 void NDUILabel::SetFontSize(unsigned int fontSize)
 {
-	fontSize = fontSize * NDDirector::DefaultDirector()->GetScaleFactor();
+	fontSize = (int)(float)fontSize * NDDirector::DefaultDirector()->GetScaleFactor();
 
 	if (m_uiFontSize != fontSize)
 	{
